* commit '63e0128e9d9a3b15c391c4ee10919eb2585b3c19': (142 commits)
  Consumer C API - create asapo string from C string
  fix AsapoConfig.cmake
  added static link flags to cmake config
  changelog
  use imported curl target
  replace add_link_options - incompatible with older cmake versions
  fix build, remove libfabric from pkg-config
  Add pkg-config files for consumer and producer APIs
  shared libraries, cmake config files
  libfabric as objects
  get rid of last warnigns
  Add pkg-config files for consumer and producer APIs
  relax some warnings, fix others
  fix makefiles
  rollback libfabric changes
  rollback libfabric changes
  fix flags
  fix file monitor
  some fixes
  fix some warnings
  ...

+++ common/cpp/include/asapo/request/request.h
@@ -28,8 +28,9 @@ class GenericRequest {
         if (timeout_ms_ == 0) {
             return false;
         }
-        uint64_t elapsed_ms = std::chrono::duration_cast<std::chrono::milliseconds>(std::chrono::system_clock::now() -
-                              created_at_).count();
+        uint64_t elapsed_ms = static_cast<uint64_t>(std::chrono::duration_cast<std::chrono::milliseconds>
+                                                    (std::chrono::system_clock::now() -
+                                                     created_at_).count());
         return elapsed_ms > timeout_ms_;
     }
