+++ tests/manual/performance_broker_receiver/getlast_broker.cpp
@@ -35,7 +35,7 @@ struct Args {
     std::string beamtime_id;
     std::string token;
     int timeout_ms;
-    int nthreads;
+    size_t nthreads;
     bool read_data;
     bool datasets;
 };
@@ -57,13 +57,13 @@ std::vector<std::thread> StartThreads(const Args& params,
                                       std::vector<int>* errors,
                                       std::vector<int>* nbuf,
                                       std::vector<int>* nfiles_total) {
-    auto exec_next = [&params, nfiles, errors, nbuf, nfiles_total](int i) {
+    auto exec_next = [&params, nfiles, errors, nbuf, nfiles_total](size_t i) {
         asapo::MessageMeta fi;
         Error err;
         auto consumer = asapo::ConsumerFactory::CreateConsumer(params.server, params.file_path, true,
-                                                             asapo::SourceCredentials{asapo::SourceType::kProcessed,
-                                                                                      params.beamtime_id, "", "",
-                                                                                      params.token}, &err);
+                        asapo::SourceCredentials{asapo::SourceType::kProcessed,
+                                                 params.beamtime_id, "", "",
+                                                 params.token}, &err);
         consumer->SetTimeout((uint64_t) params.timeout_ms);
         asapo::MessageData data;
 
@@ -111,7 +111,7 @@ std::vector<std::thread> StartThreads(const Args& params,
     };
 
     std::vector<std::thread> threads;
-    for (int i = 0; i < params.nthreads; i++) {
+    for (size_t i = 0; i < params.nthreads; i++) {
         threads.emplace_back(std::thread(exec_next, i));
     }
     return threads;
@@ -138,12 +138,12 @@ int ReadAllData(const Args& params, uint64_t* duration_ms, int* nerrors, int* nb
 
     system_clock::time_point t2 = system_clock::now();
     auto duration_read = std::chrono::duration_cast<std::chrono::milliseconds>(t2 - t1);
-    *duration_ms = duration_read.count();
+    *duration_ms = static_cast<uint64_t>(duration_read.count());
 
     // The following two loops will check if all threads that processed some data were using the same network type
     {
-        int firstThreadThatActuallyProcessedData = 0;
-        for (int i = 0; i < params.nthreads; i++) {
+        size_t firstThreadThatActuallyProcessedData = 0;
+        for (size_t i = 0; i < params.nthreads; i++) {
             if (nfiles[i] > 0) {
                 firstThreadThatActuallyProcessedData = i;
                 break;
@@ -151,7 +151,7 @@ int ReadAllData(const Args& params, uint64_t* duration_ms, int* nerrors, int* nb
         }
 
         *connection_type = connection_types[firstThreadThatActuallyProcessedData];
-        for (int i = 0; i < params.nthreads; i++) {
+        for (size_t i = 0; i < params.nthreads; i++) {
             if (*connection_type != connection_types[i] && nfiles[i] > 0) {
                 // The output will look like this:
                 // ERROR thread[0](processed 5 files) connection type is 'No connection' but thread[1](processed 3 files) is 'TCP'
@@ -181,7 +181,7 @@ int main(int argc, char* argv[]) {
     params.server = std::string{argv[1]};
     params.file_path = std::string{argv[2]};
     params.beamtime_id = std::string{argv[3]};
-    params.nthreads = atoi(argv[4]);
+    params.nthreads = static_cast<size_t>(atoi(argv[4]));
     params.token = std::string{argv[5]};
     params.timeout_ms = atoi(argv[6]);
     params.read_data = atoi(argv[7]) != 1;
@@ -203,7 +203,8 @@ int main(int argc, char* argv[]) {
     }
     std::cout << "Errors : " << nerrors << std::endl;
     std::cout << "Elapsed : " << duration_ms << "ms" << std::endl;
-    std::cout << "Rate : " << 1000.0f * nfiles / (duration_ms) << " Hz" << std::endl;
+    std::cout << "Rate : " << 1000.0f * static_cast<float>(nfiles) / (static_cast<float>
+              (duration_ms)) << " Hz" << std::endl;
 
     std::cout << "Using connection type: " << ConnectionTypeToString(connectionType) << std::endl;
